Code refactoring for message package
Code refactoring for message package
