+++ packages/message/src/property/index.ts
@@ -1,0 +1,3 @@
+export * from './hashProperty';
+export * from './objectToProperty';
+export * from './propertyObject';
