+++ packages/message/src/property/objectToProperty.ts
@@ -1,0 +1,11 @@
+import { PropertyType, Property } from '../type';
+
+export function objectToProperty(object: {
+  [key: string]: string;
+}): Property[] {
+  return Object.keys(object).map((key) => ({
+    type: PropertyType.Raw,
+    key,
+    value: object[key],
+  }));
+}
