+++ packages/message/src/utils/hash.ts
@@ -1,0 +1,5 @@
+import sha256 from 'crypto-js/sha256';
+
+export function hash(msg: string): string {
+  return sha256(msg).toString();
+}
