+++ packages/message/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './hash';
+export * from './utf8ToBase64';
