+++ packages/message/src/utils/utf8ToBase64.ts
@@ -1,0 +1,3 @@
+export function utf8ToBase64(str: string): string {
+  return Buffer.from(encodeURIComponent(str), 'binary').toString('base64');
+}
